Add two really bad gaussian blur programs
First is a version that works on 30x30 but does not need any additional
features. Second is a prototype version that works on much larger image
but needs additional features.

For both:
- Do not work on the edges
- Some values are missing
- Probably suboptimal

+++ lib/assembler.h
@@ -28,7 +28,7 @@ namespace Assembler
     using Ast = std::unordered_map<BlockId, std::vector<Node>>;
 
     /// Parsed settings, part of the AST
-    using SettingMap = std::unordered_map<std::string, std::vector<uint8_t>>;
+    using SettingMap = std::unordered_map<std::string, std::vector<uint16_t>>;
 
     /// Preprocess input into output
     void preprocess(std::istream& input, std::ostream& output);
